requirements.txt
Django==5.0
gunicorn
whitenoise
pytz
numpy
pandas
scikit-learn
faiss-cpu
djangorestframework
django-cors-headers
torch==2.4.1
gradio_client
requests==2.32.3